+++ src/main/kotlin/hr/eduwalk/data/database/table/QuestionTable.kt
@@ -5,7 +5,7 @@ import org.jetbrains.exposed.sql.Table
 
 object QuestionTable : Table() {
 
-    val id = integer(name = "id").autoIncrement()
+    val id = long(name = "id").autoIncrement()
     val questionText = text(name = "question_text")
     val answer0 = text(name = "answer0")
     val answer1 = text(name = "answer1")
@@ -13,7 +13,7 @@ object QuestionTable : Table() {
     val answer3 = text(name = "answer3").nullable()
     val answer4 = text(name = "answer4").nullable()
     val correctAnswer = text(name = "correct_answer")
-    val locationId = integer(name = "location_id")
+    val locationId = long(name = "location_id")
         .index(isUnique = false)
         .references(
             ref = LocationTable.id,
