+++ src/main/kotlin/hr/eduwalk/domain/interfaces/ILocationScoreDao.kt
@@ -6,5 +6,5 @@ import hr.eduwalk.domain.model.ServiceResult
 interface ILocationScoreDao {
     suspend fun insertLocationScore(locationScore: LocationScore): ServiceResult<Unit>
     suspend fun updateLocationScore(locationScore: LocationScore): ServiceResult<Boolean>
-    suspend fun getLocationScoreForUser(locationId: Int, userId: String): ServiceResult<LocationScore>
+    suspend fun getLocationScoreForUser(locationId: Long, userId: String): ServiceResult<LocationScore>
 }
